osan 4 koodit

+++ osa4/bloglist/controllers/blogs.js
@@ -1,8 +1,14 @@
 const blogsRouter = require('express').Router()
+const tokenExtractor = require('../middleware/tokenextractor')
+const userExtractor = require('../middleware/userextractor')
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {
+    username: 1,
+    name: 1,
+    id: 1,
+  })
   res.status(200).json(blogs)
 })
 
@@ -13,13 +19,22 @@ blogsRouter.get('/:id', async (req, res) => {
   else res.status(404).end()
 })
 
-blogsRouter.post('/', async (req, res) => {
-  const newBlog = new Blog(req.body)
+blogsRouter.post('/', tokenExtractor, userExtractor, async (req, res) => {
+  const newBlog = new Blog({ ...req.body, user: req.user._id })
   const result = await newBlog.save()
   res.status(201).json(result)
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', tokenExtractor, userExtractor, async (req, res) => {
+  const blogToRemove = await Blog.findById(req.params.id)
+  if (
+    !req.user._id ||
+    !blogToRemove.user.toString() ||
+    blogToRemove.user.toString() !== req.user.id.toString()
+  ) {
+    return res.status(403).json({ error: 'not authorized to remove that blog' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
